Keep Carbon status expectations type-exact
Convert the named Carbon eventNotHandledErr constant to Int32 in both native expectations so XCTest compares one concrete OSStatus-compatible type.

Constraint: Preserve the production API and named Carbon status semantics.
Rejected: Use a numeric literal or broaden the service return type.
Confidence: High; this directly resolves both generic-parameter diagnostics.
Scope-risk: Two native-test expected values only.
Reversibility: Revert this GREEN and its immediate RED parent together.
Directive: Keep unknown signatures and identifiers explicitly not handled.
Tested: Focused status-type contract passes; diff check passes.
Not-tested: Full WSL aggregate and macOS test build after repair.
Related: controlled GitHub Actions run 29551936506.

+++ PrivatePresenterAppTests/CarbonHotKeyServiceTests.swift
@@ -98,13 +98,13 @@ final class CarbonHotKeyServiceTests: XCTestCase {
 
         XCTAssertEqual(
             harness.service.receiveForTesting(identifier: .init(signature: 0, id: 1)),
-            eventNotHandledErr
+            Int32(eventNotHandledErr)
         )
         XCTAssertEqual(
             harness.service.receiveForTesting(
                 identifier: .init(signature: ProductHotKeyIdentifier.signature, id: 99)
             ),
-            eventNotHandledErr
+            Int32(eventNotHandledErr)
         )
         XCTAssertTrue(harness.commands.isEmpty)
     }
